+++ .gitignore
@@ -9,3 +9,4 @@ db.sqlite3
 
 .vscode/
 
+*.env
