+++ frontend/lib/feature/book_search/search_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
-import 'package:book_chat/model/model_book.dart';
+import 'package:book_chat/model/book_model.dart';
 import 'package:book_chat/feature/book_info/book_info_screen.dart';
 
 class SearchScreen extends StatefulWidget {
