+++ frontend/lib/feature/sign_up/signup_screen.dart
@@ -1,6 +1,7 @@
-import 'package:book_chat/feature/sign_up/signup_dto.dart';
-import 'package:book_chat/feature/sign_up/signup_notifier.dart';
-import 'package:book_chat/feature/sign_up/signup_repository.dart';
+import 'package:book_chat/common/repository/api_repository.dart';
+import 'package:book_chat/dto/signup_dto.dart';
+import 'package:book_chat/feature/sign_up/providers/signup_notifier.dart';
+import 'package:book_chat/data/sign_up/signup_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart' as http;
