+++ frontend/lib/model/book_model.dart
