+++ frontend/pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
+  flutter_dotenv: ^5.2.1
 
 dev_dependencies:
   flutter_test:
@@ -78,6 +79,7 @@ flutter:
   assets:
     - images/
     - images/nousb.png
+    - .env
   #  - images/beasts_of_a_little_land.png
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
